chore: refactor css and structure
