+++ src/styles/LoadingSpinner.css
