+++ src/styles/TablePage.css
@@ -1,8 +1,3 @@
-:root {
-  --danger-color: #ff4d4f;
-  --error-bg: #fff1f0;
-}
-
 h1 {
   font-size: 1.5rem;
   margin: 1rem 0;
@@ -28,20 +23,24 @@ h1 {
   gap: 10px;
 }
 
+main.container {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+}
+
 /* Table wrapper for overflow handling */
 .table-wrapper {
   overflow-x: auto;
-  margin-bottom: 1rem;
   font-size: 0.9rem;
+  display: flex;
+  flex-direction: column;
+  flex: 1;
 }
 
-/* Truncation for body text in table and cards could be added 
-.body-text {
-  max-width: 300px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-} */
+.table-view {
+  flex-grow: 1;
+}
 
 /* Sortable column headers */
 th.sortable {
@@ -129,26 +128,38 @@ th.sortable {
 
 /* Pagination styles */
 .pagination {
+  position: sticky;
+  bottom: 0;
   display: flex;
   justify-content: space-between;
   align-items: center;
   gap: 1rem;
-  flex-wrap: wrap;
+  z-index: 10;
+  margin: 1rem 0;
 }
 
 .pagination div {
   display: flex;
-  flex-wrap: wrap;
+  font-size: 0.7rem;
+  align-items: center;
+}
+
+.pagination span {
+  min-width: 100px;
 }
 
 .pagination button {
   margin-right: 0.3rem;
+  min-width: 80px;
 }
 
 .pagination select {
   padding: 0.3rem;
   font-size: 0.85rem;
   margin: 0;
+  font-size: 0.7rem;
+  height: 2rem;
+  min-width: 100px;
 }
 
 .empty-state {
@@ -180,6 +191,10 @@ th.sortable {
   }
   .sort-buttons {
     display: flex;
+  }
+
+  .pagination {
+    flex-direction: column;
   }
 }
